move get_locale init to ext
